+++ src/main/kotlin/cloud/glitchdev/rfu/feature/mob/JawbusHardMode.kt
@@ -3,6 +3,7 @@ package cloud.glitchdev.rfu.feature.mob
 import cloud.glitchdev.rfu.RiccioFishingUtils.mc
 import cloud.glitchdev.rfu.config.categories.LavaFishing
 import cloud.glitchdev.rfu.events.managers.ChatEvents.registerGameEvent
+import cloud.glitchdev.rfu.events.managers.ShutdownEvents
 import cloud.glitchdev.rfu.feature.Feature
 import cloud.glitchdev.rfu.feature.RFUFeature
 import cloud.glitchdev.rfu.gui.window.DeadWindow
@@ -24,6 +25,7 @@ object JawbusHardMode : Feature {
                     delay(2000)
 
                     mc.execute {
+                        ShutdownEvents.runTasks(mc)
                         mc.emergencySaveAndCrash(CrashReport("You died to lord jawbus.", RuntimeException("Skill Issue")))
                     }
                 }