+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# v1.5.1 - ???
+# v1.5.1 - Achievement Patches
 
 ### Features:
 - Added 7 more achievements